use unsigned long int for numfields and for indexes

+++ include/dbi/dbd.h
@@ -55,9 +55,9 @@ int dbd_conn_quote_string(dbi_conn_t *conn, const char *orig, char *dest);
 
 /* _DBD_* DRIVER AUTHORS HELPER FUNCTIONS */
 dbi_result_t *_dbd_result_create(dbi_conn_t *conn, void *handle, unsigned long long numrows_matched, unsigned long long numrows_affected);
-void _dbd_result_set_numfields(dbi_result_t *result, unsigned short numfields);
+void _dbd_result_set_numfields(dbi_result_t *result, unsigned long numfields);
 void _dbd_result_add_field(dbi_result_t *result, unsigned short idx, char *name, unsigned short type, unsigned int attribs);
-dbi_row_t *_dbd_row_allocate(unsigned short numfields);
+dbi_row_t *_dbd_row_allocate(unsigned long numfields);
 void _dbd_row_finalize(dbi_result_t *result, dbi_row_t *row, unsigned long long idx);
 void _dbd_internal_error_handler(dbi_conn_t *conn, const char *errmsg, const int errno);
 dbi_result_t *_dbd_result_create_from_stringarray(dbi_conn_t *conn, unsigned long long numrows_matched, const char **stringarray);
